Adapt to HAF op_type_id column removal from operation ID encoding
Update btracker for the new hafd.operation_id(block_num, op_pos)
encoding which no longer includes op_type_id. The op_type_id is now
stored as a separate column in hafd.operations instead of being
encoded into the operation ID.

Changes:
- Use 2-arg hafd.operation_id(block_num, op_pos) in mock insert
- Insert op_type_id as explicit column in mock operations
- Update mock fixture op_pos values to be unique per block
- Use dynamic hafd.operation_id() for hardfork_vop_id instead of
  hardcoded value
- Update all pattern files with new operation_id values
- Update endpoint type descriptions
- Capture fill-db container logs on CI failure
- Fix publish_images for tag pipelines

+++ tests/mocks/sql/types.sql
@@ -31,7 +31,7 @@ SET ROLE btracker_owner;
 DROP TYPE IF EXISTS btracker_backend.operation_type CASCADE;
 CREATE TYPE btracker_backend.operation_type AS (
     block_num INT,
-    op_type_id INT,
+    op_type_id SMALLINT,
     op_pos INT,
     trx_in_block INT,
     body JSON
